requirements.txt
playwright==1.40.0
pytest-playwright==0.4.3
pytest==7.4.0
psycopg2-binary==2.9.7
pymysql==1.1.0